refactor(data): promote core mutators + merge helpers out of internals
Move the mutator surface plugins already depend on out of
`src/data/internals/` into public `@/data/` modules — no shims, the old
import paths are gone:

- kernelMutators.ts -> src/data/mutators.ts (createChild, merge, setContent,
  revealChildren, ...). It imported only ./blockMerge internally and nothing
  in internals/ imports it back, so the whole module moves cleanly.
- blockMerge.ts -> src/data/blockMerge.ts (mergeBlocksInTx, ContentStrategy)
- mergeProperties.ts -> src/data/mergeProperties.ts

Repoint all importers (repo, kernelDataExtension, typeExtraction,
exampleExtensions, defaultShortcuts, paste, merge-blocks, alias) to the
public paths. `tx.run` takes the mutator object, so these genuinely need
the export; promoting the module is the clean way to give it to them.

Part of #139 (promotion-first pass; no lint rule).
