+++ src/data/mergeProperties.test.ts
