+++ src/data/mergeProperties.ts
